Split <StopTableRow> into smaller subcomponents
