+++ ui/src/components/stop-registry/search/StopTableRow/StopRowTdProps.ts
@@ -1,0 +1,6 @@
+import { StopSearchRow } from '../../../../hooks';
+
+export type StopRowTdProps = {
+  readonly className?: string;
+  readonly stop: StopSearchRow;
+};
