+++ ui/src/components/stop-registry/search/StopTableRow/index.ts
@@ -1,0 +1,1 @@
+export * from './StopTableRow';
