+++ ui/src/components/stop-registry/search/index.ts
@@ -1,4 +1,4 @@
 export * from './StopSearchBar';
 export * from './StopSearchResultList';
 export * from './StopSearchResultsPage';
-export * from './StopTableRow';
+export * from './StopTableRow/StopTableRow';
